+++ Tests/Tests/Collections/MemoryAllocator/V3/HashSet.cs
@@ -4,14 +4,14 @@ namespace ME.ECS.Tests.MemoryAllocator.V3.Collections {
     using UnityEngine;
     using ME.ECS.Collections.MemoryAllocator;
 
-    public class List {
+    public class HashSet {
 
         [Test]
         public void Initialize() {
 
             var allocator = Base.GetAllocator(10);
 
-            var list = new List<Vector3>(ref allocator, 10);
+            var list = new HashSet<Vector3>(ref allocator, 10);
             Assert.IsTrue(list.isCreated);
             list.Dispose(ref allocator);
 
@@ -24,36 +24,13 @@ namespace ME.ECS.Tests.MemoryAllocator.V3.Collections {
 
             var allocator = Base.GetAllocator(10);
 
-            var list = new List<Vector3>(ref allocator, 10);
+            var list = new HashSet<Vector3>(ref allocator, 10);
             for (int i = 0; i < 100; ++i) {
 
                 list.Add(ref allocator, new Vector3(i, i, i));
                 
             }
             
-            Assert.IsTrue(list.Count == 100);
-
-            allocator.Dispose();
-
-        }
-
-        [Test]
-        public void RemoveAt() {
-
-            var allocator = Base.GetAllocator(10);
-
-            var list = new List<Vector3>(ref allocator, 10);
-            for (int i = 0; i < 100; ++i) {
-
-                list.Add(ref allocator, new Vector3(i, i, i));
-                
-            }
-            
-            Assert.IsTrue(list.Count == 100);
-            Assert.IsTrue(list.RemoveAt(ref allocator, 50));
-            Assert.IsTrue(list.Count == 99);
-            Assert.IsFalse(list.RemoveAt(ref allocator, 100));
-            list.Add(ref allocator, new Vector3());
             Assert.IsTrue(list.Count == 100);
 
             allocator.Dispose();
@@ -65,7 +42,7 @@ namespace ME.ECS.Tests.MemoryAllocator.V3.Collections {
 
             var allocator = Base.GetAllocator(10);
 
-            var list = new List<Vector3>(ref allocator, 10);
+            var list = new HashSet<Vector3>(ref allocator, 10);
             for (int i = 0; i < 100; ++i) {
 
                 list.Add(ref allocator, new Vector3(i, i, i));
@@ -79,34 +56,11 @@ namespace ME.ECS.Tests.MemoryAllocator.V3.Collections {
         }
 
         [Test]
-        public void RemoveAtFast() {
-
-            var allocator = Base.GetAllocator(10);
-
-            var list = new List<Vector3>(ref allocator, 10);
-            for (int i = 0; i < 100; ++i) {
-
-                list.Add(ref allocator, new Vector3(i, i, i));
-                
-            }
-            
-            Assert.IsTrue(list.Count == 100);
-            Assert.IsTrue(list.RemoveAtFast(ref allocator, 50));
-            Assert.IsTrue(list.Count == 99);
-            Assert.IsFalse(list.RemoveAtFast(ref allocator, 100));
-            list.Add(ref allocator, new Vector3());
-            Assert.IsTrue(list.Count == 100);
-
-            allocator.Dispose();
-
-        }
-
-        [Test]
         public void Remove() {
 
             var allocator = Base.GetAllocator(10);
 
-            var list = new List<Vector3>(ref allocator, 10);
+            var list = new HashSet<Vector3>(ref allocator, 10);
             for (int i = 0; i < 100; ++i) {
 
                 list.Add(ref allocator, new Vector3(i, i, i));
@@ -126,7 +80,7 @@ namespace ME.ECS.Tests.MemoryAllocator.V3.Collections {
 
             var allocator = Base.GetAllocator(10);
 
-            var list = new List<Vector3>(ref allocator, 10);
+            var list = new HashSet<Vector3>(ref allocator, 10);
             for (int i = 0; i < 100; ++i) {
 
                 list.Add(ref allocator, new Vector3(i, i, i));